requirements.txt
colorama
datasets
fpdf
openai
langchain
langchain-community
langchain-openai
langchain-experimental
langchain-neo4j
neo4j
python-dotenv
pypdf
pydantic
pydantic_core
streamlit
json-repair
langchain-nomic
rdflib
langchain-google-genai
rapidfuzz
numpy
PyMuPDF